Little tweaking As a note, I removed Livewire from the Dashboard view by default, mainly because an Index may not need Livewire at all (depends on implementation).
However, Alpine will be kept for the Sidebar stuff. This may sound problematic, since you need to load Livewire before Alpine to evade some reactivity
issues, however, Blade templates also have a rendering order. Children views are parsed before the parent layout, so if we call for Livewire usage
in a children view, Livewire will be added first to the stack (including 'scripts' and 'styles'). After that, the layout view will be called, so
we'll get Alpine stacked "after" livewire.